move bounded series iterator
